[Setup] DRY out header & footer instantiation to hooks
- this commit isn't specifically needed by this PR, but *will* be needed once we implement a custom body renderer
